Day 2 Features (#5)
* Chording

Pressing enter on a revealed square may reveal adjacent unflagged squares if the number of flags around the square is equal to its number of surrounding water tiles (mines)

* Ensure Safe First Reveal

The first square revealed is guaranteed to have 0 mines surrounding it.

* Distinguish Revealed Squares

Can no longer flag squares with numbers/revealed squares. Marked revealed squares as a distinct color

* Lives and Loss

Added 3 hearts in top right corner that get depleted when the player clicks on mines. If all lives are lost, the game is lost and the player can no longer do anything