added multiple changes that all work:
win condition checker: created new gamestate when all non mined tiles are revealed.

made it possible to win: added minimap opacity if the player collides with it so that you can see under it.

added dev cheatcode ability for eassier playtesting: if press y, 21x21 area around the player revealed. should comment this out or add a toggle in the future, but for now helps speed up dev testing.